requirements.full.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.0
assemblyai==0.34.0
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
cloudscraper==1.2.71
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
cssselect==1.2.0
cssutils==2.11.1
curl_cffi==0.7.3
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==4.4.2
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
g2pkk==0.1.2
g4f==0.3.3.2
geckodriver-autoinstaller==0.1.0
grpcio==1.66.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
inflect==7.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsoncodable==0.1.7
jsonlines==1.2.0
jsonpickle==3.3.0
k-selenium-cookies==0.0.4
kiwisolver==1.4.7
kproxy==0.0.1
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
msvc_runtime==14.40.33807
multidict==6.1.0
murmurhash==1.0.10
networkx==2.8.8
nltk==3.9.1
noraise==0.0.16
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==1.5.3
Pillow==9.5.0
platformdirs==4.3.6
pooch==1.8.2
premailer==3.10.0
preshed==3.0.9
prettytable==3.11.0
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
pypinyin==0.53.0
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.2
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
selenium-browser==0.0.15
selenium-firefox==2.0.8
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
srt==3.5.3
srt_equalizer==0.1.10
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.20.1
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
trainer==0.0.36
transformers==4.45.2
trio==0.26.2
trio-websocket==0.11.1
TTS==0.22.0
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.6
undetected-chromedriver==3.5.5
Unidecode==1.3.8
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
xpath-utils==0.0.3
yagmail==0.15.293
yarl==1.14.0
